requirements.txt
streamlit
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
transformers
pillow
geopy
folium
streamlit-folium